requirements.txt
-i https://pypi.org/simple
blinker==1.9.0; python_version >= '3.9'
boto3==1.40.30; python_version >= '3.9'
botocore==1.40.30; python_version >= '3.9'
certifi==2025.8.3; python_version >= '3.7'
cffi==2.0.0; python_version >= '3.9'
charset-normalizer==3.4.3; python_version >= '3.7'
click==8.2.1; python_version >= '3.10'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
copyparty==1.19.9; python_version >= '3.3'
cryptography==45.0.7; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
distlib==0.4.0
dnspython==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
filelock==3.19.1; python_version >= '3.9'
flask==3.1.2; python_version >= '3.9'
flask-cors==6.0.1; python_version >= '3.9' and python_version < '4.0'
flask-pymongo==2.3.0
google-cloud-storage==2.18.0; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
packaging==25.0; python_version >= '3.8'
pbr==7.0.1; python_version >= '2.6'
pipenv==2025.0.4; python_version >= '3.9'
pipenv-to-requirements==0.9.0
platformdirs==4.4.0; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pycparser==2.23; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
pymongo[srv]==3.12.0
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.1.1; python_version >= '3.9'
python-json-logger==3.3.0; python_version >= '3.8'
redis==6.4.0; python_version >= '3.9'
requests==2.32.5; python_version >= '3.9'
s3transfer==0.14.0; python_version >= '3.9'
setuptools==80.9.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
urllib3==2.5.0; python_version >= '3.9'
virtualenv==20.34.0; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
